prepWork
In preparation to conduct dialect improvement sprints for the DataOne
dialects, I decided to prepare systematically and came up with the DISC
analysis, Dialect Improvement Sprint Coordinator.

The DISC spreadsheet utilizes the dialect and recommendation tables
that can be generated with /crosswalks/tables/xsl to determine which of
the concepts in AllCrosswalks.xml were included in the 11 DUG dashboard
ready recommendations and which of those concepts in each DataOne
dialect might need a definition.