EML_DISC2
It looks like there are about 40 elements that appear in LTER/KNB EML
records that are not defined. Most of these seem to be dealing with
parts of document structure, and metadata describing the data table
entity/attribute values and meanings. Most of these things are not in
any of the recommendations I have selected for the DUG reports, nor for
the ESIP metadata improvement lab. Many of them look like new (to us)
concepts. Setting this dialect improvement sprint coordinator off to
the side for a rainy day