newIdeas
Would like to provide actionable output that directly links to the
guidance wikis on the esipfed servers. Currently needs to remove blanks
from list before ready to share. This is also an issue with the
incomplete concepts visualization, may need some vba for this as the
only potential answers I’ve located are a manual process (not
sufficient) or a short vba code (sufficient but not understood….)