incorporatingRDA-CISL
Spoke with Bob Dattore about the Metadata Lab. He said he would like to
use some of his metadata from RDA at CISL. Since the NCAR-DSET contract
gave me the opportunity to incorporate this dialect into AllCrosswalks
this is a low-effort ask, and a great opportunity to check back in with
NCAR with the increased analysis ability.